+++ apps/app/src/test/scala/com/daml/network/integration/tests/GlobalDomainMigrationIntegrationTest.scala
@@ -438,10 +438,22 @@ class GlobalDomainMigrationIntegrationTest
             }
           }
 
-          startAllSync(
-            sv1ScanLocalBackend,
-            sv1ValidatorLocalBackend,
+          // TODO(#9977)
+          loggerFactory.assertLogsSeq(SuppressionRule.LevelAndAbove(Level.ERROR))(
+            startAllSync(
+              sv1ScanLocalBackend,
+              sv1ValidatorLocalBackend,
+            ),
+            (
+                entries =>
+                  forAll(entries) {
+                    _.errorMessage should include(
+                      "Unexpected coin create event"
+                    )
+                  }
+            ),
           )
+
           sv1LocalBackend.getSvcInfo().svcRules.payload.members.size() shouldBe 4
 
           clue("Old wallet balance is recorded") {